+++ test/cbmc/proofs/TCP/prvHandleListen/Makefile.json
@@ -10,7 +10,7 @@
     "$(ENTRY)_harness.goto",
     "$(FREERTOS_PLUS_TCP)/test/cbmc/stubs/cbmc.goto",
     "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_State_Handling.goto",
-    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_State_Handling_IPV4.goto"
+    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_State_Handling_IPv4.goto"
   ],
   "DEF":
   [
