+++ test/cbmc/proofs/TCP/prvHandleListen_IPv6/Makefile.json
@@ -12,7 +12,7 @@
     "$(ENTRY)_harness.goto",
     "$(FREERTOS_PLUS_TCP)/test/cbmc/stubs/cbmc.goto",
     "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_State_Handling.goto",
-    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_State_Handling_IPV6.goto"
+    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_State_Handling_IPv6.goto"
   ],
   "DEF":
   [
