+++ test/cbmc/proofs/parsing/ProcessReceivedTCPPacket_IPv6/Makefile.json
@@ -11,9 +11,9 @@
     "$(ENTRY)_harness.goto",
     "$(FREERTOS_PLUS_TCP)/test/cbmc/stubs/cbmc.goto",
     "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_IP.goto",
-    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_IP_IPV6.goto",
+    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_IP_IPv6.goto",
     "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_Transmission.goto",
-    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_Transmission_IPV6.goto"
+    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_TCP_Transmission_IPv6.goto"
   ],
   "INSTFLAGS":
   [
