+++ test/unit-test/CMakeLists.txt
@@ -302,13 +302,13 @@ include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_Reception/ut.cmake )
 include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_IP/ut.cmake )
 include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_IP_DiffConfig/ut.cmake )
 include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_State_Handling/ut.cmake )
-include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_State_Handling_IPV4/ut.cmake )
-include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_State_Handling_IPV6/ut.cmake )
+include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_State_Handling_IPv4/ut.cmake )
+include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_State_Handling_IPv6/ut.cmake )
 include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_Transmission/ut.cmake )
-include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_Transmission_IPV6/ut.cmake )
+include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_Transmission_IPv6/ut.cmake )
 include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_Utils/ut.cmake )
-include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_Utils_IPV6/ut.cmake )
-include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_Utils_IPV6_ConfigLowTCPMSS/ut.cmake )
+include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_Utils_IPv6/ut.cmake )
+include( ${UNIT_TEST_DIR}/FreeRTOS_TCP_Utils_IPv6_ConfigLowTCPMSS/ut.cmake )
 include( ${UNIT_TEST_DIR}/FreeRTOS_IPv6/ut.cmake )
 include( ${UNIT_TEST_DIR}/FreeRTOS_IPv6_ConfigDriverCheckChecksum/ut.cmake )
 include( ${UNIT_TEST_DIR}/FreeRTOS_Sockets_IPv6/ut.cmake )
@@ -348,14 +348,14 @@ add_custom_target( coverage
     FreeRTOS_Tiny_TCP_utest
     FreeRTOS_TCP_Reception_utest
     FreeRTOS_TCP_State_Handling_utest
-    FreeRTOS_TCP_State_Handling_IPV4_utest
-    FreeRTOS_TCP_State_Handling_IPV6_utest
+    FreeRTOS_TCP_State_Handling_IPv4_utest
+    FreeRTOS_TCP_State_Handling_IPv6_utest
     FreeRTOS_TCP_Transmission
-    FreeRTOS_TCP_Transmission_IPV6
+    FreeRTOS_TCP_Transmission_IPv6
     FreeRTOS_TCP_IP
     FreeRTOS_TCP_Utils
-    FreeRTOS_TCP_Utils_IPV6_utest
-    FreeRTOS_TCP_Utils_IPV6_ConfigLowTCPMSS_utest
+    FreeRTOS_TCP_Utils_IPv6_utest
+    FreeRTOS_TCP_Utils_IPv6_ConfigLowTCPMSS_utest
     FreeRTOS_Sockets_GenericAPI_utest
     FreeRTOS_Sockets_privates_utest
     FreeRTOS_Sockets_TCP_API_utest
