+++ test/unit-test/FreeRTOS_TCP_IP_DiffConfig/ut.cmake
@@ -52,7 +52,7 @@ set(real_source_files "")
 # list the files you would like to test here
 list(APPEND real_source_files
             ${CMAKE_BINARY_DIR}/Annexed_TCP_Sources/FreeRTOS_TCP_IP.c
-            ${CMAKE_BINARY_DIR}/Annexed_TCP_Sources/FreeRTOS_TCP_IP_IPV4.c
+            ${CMAKE_BINARY_DIR}/Annexed_TCP_Sources/FreeRTOS_TCP_IP_IPv4.c
             ${MODULE_ROOT_DIR}/test/unit-test/${project_name}/${project_name}_stubs.c
 	)
