+++ test/unit-test/FreeRTOS_TCP_State_Handling_IPv4/ut.cmake
@@ -2,7 +2,7 @@
 include( ${MODULE_ROOT_DIR}/test/unit-test/TCPFilePaths.cmake )
 
 # ====================  Define your project name (edit) ========================
-set( project_name "FreeRTOS_TCP_State_Handling_IPV4" )
+set( project_name "FreeRTOS_TCP_State_Handling_IPv4" )
 message( STATUS "${project_name}" )
 
 # =====================  Create your mock here  (edit)  ========================
