+++ test/unit-test/FreeRTOS_TCP_Utils_IPv6/ut.cmake
@@ -2,7 +2,7 @@
 include( ${MODULE_ROOT_DIR}/test/unit-test/TCPFilePaths.cmake )
 
 # ====================  Define your project name (edit) ========================
-set( project_name "FreeRTOS_TCP_State_Handling_IPV6" )
+set( project_name "FreeRTOS_TCP_Utils_IPv6" )
 message( STATUS "${project_name}" )
 
 # =====================  Create your mock here  (edit)  ========================
@@ -10,16 +10,9 @@ set(mock_list "")
 
 # list the files to mock here
 list(APPEND mock_list
-            "${MODULE_ROOT_DIR}/test/FreeRTOS-Kernel/include/task.h"
-            "${MODULE_ROOT_DIR}/test/FreeRTOS-Kernel/include/list.h"
-            "${MODULE_ROOT_DIR}/test/FreeRTOS-Kernel/include/queue.h"
-            "${MODULE_ROOT_DIR}/test/FreeRTOS-Kernel/include/event_groups.h"
             "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_IP.h"
-            "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_IP_Private.h"
             "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_Sockets.h"
-            "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_TCP_Transmission.h"
-            "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_TCP_Utils.h"
-            "${MODULE_ROOT_DIR}/test/unit-test/${project_name}/TCP_State_Handling_IPV6_list_macros.h"
+            "${CMAKE_BINARY_DIR}/Annexed_TCP/FreeRTOS_Routing.h"
         )
 
 set(mock_include_list "")
@@ -29,8 +22,8 @@ list(APPEND mock_include_list
             ${TCP_INCLUDE_DIRS}
             ${MODULE_ROOT_DIR}/test/FreeRTOS-Kernel/include
             ${MODULE_ROOT_DIR}/test/FreeRTOS-Kernel/portable/ThirdParty/GCC/Posix
-            ${MODULE_ROOT_DIR}/test/unit-test/${project_name}
             ${MODULE_ROOT_DIR}/test/unit-test/ConfigFiles
+            ${MODULE_ROOT_DIR}/test/unit-test/${project_name}
         )
 
 set(mock_define_list "")
